created form 10

+++ common/modules/rpmes/controllers/ProjectProblemController.php
@@ -41,12 +41,16 @@ class ProjectProblemController extends Controller
             ],
             'access' => [
                 'class' => AccessControl::className(),
-                'only' => ['index', 'create', 'update', 'generate', 'delete'],
                 'rules' => [
+                    [
+                        'actions' => ['index'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
                     [
                         'actions' => ['index', 'create', 'update', 'generate', 'delete'],
                         'allow' => true,
-                        'roles' => ['@'],
+                        'roles' => ['Administrator', 'SuperAdministrator'],
                     ],
                 ],
             ],
@@ -104,6 +108,7 @@ class ProjectProblemController extends Controller
             $model->submitted_by = Yii::$app->user->id;
             $model->date_submitted = date('Y-m-d H:i:s');
             $model->save();
+
             \Yii::$app->getSession()->setFlash('success', 'Record Saved');
             return $this->redirect(['index']);
         }
@@ -160,9 +165,12 @@ class ProjectProblemController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-        \Yii::$app->getSession()->setFlash('success', 'Record Deleted');
-        return $this->redirect(['index']);
+        if(Yii::$app->request->post())
+        {
+            $this->findModel($id)->delete();
+            \Yii::$app->getSession()->setFlash('success', 'Record Deleted');
+            return $this->redirect(['index']);
+        }
     }
 
     public function actionGenerate()
@@ -338,274 +346,5 @@ class ProjectProblemController extends Controller
         }
 
         throw new NotFoundHttpException('The requested page does not exist.');
-    }
-
-    public function actionPrintFormEleven($year,$quarter)
-    {
-        $model = [];
-        $model['year'] = $year;
-        $model['quarter'] = $quarter;
-
-        //echo "<pre>"; print_r($model['year']); exit;
-
-        $financials = ProjectTarget::find()->where(['target_type' => 'Financial', 'year' => $model['year']])->createCommand()->getRawSql();
-        
-        $financialTotal = 'IF(project.data_type = "Cumulative",
-        IF(COALESCE(financials.q4, 0) <= 0,
-            IF(COALESCE(financials.q3, 0) <= 0,
-                IF(COALESCE(financials.q2, 0) <= 0,
-                     COALESCE(financials.q1, 0)
-                        , COALESCE(financials.q2, 0)
-                        )
-                    , COALESCE(financials.q3, 0)
-                    )
-                , COALESCE(financials.q4, 0))
-            ,   
-            COALESCE(financials.q1, 0) +
-            COALESCE(financials.q2, 0) +
-            COALESCE(financials.q3, 0) +
-            COALESCE(financials.q4, 0)
-            )';
-
-            $regionTitles = ProjectRegion::find()
-            ->select(['project_id', 'GROUP_CONCAT(DISTINCT tblregion.abbreviation ORDER BY tblregion.abbreviation ASC SEPARATOR ", ") as title'])
-            ->leftJoin('tblregion', 'tblregion.region_c = project_region.region_id')
-            ->leftJoin('project', 'project.id = project_region.project_id')
-            ->where(['project.draft' => 'No'])
-            ->groupBy(['project_region.project_id'])
-            ->createCommand()->getRawSql();
-
-            $provinceTitles = ProjectProvince::find()
-            ->select(['project_id', 'GROUP_CONCAT(DISTINCT tblprovince.province_m ORDER BY tblprovince.province_m ASC SEPARATOR ", ") as title'])
-            ->leftJoin('tblprovince', 'tblprovince.province_c = project_province.province_id')
-            ->leftJoin('project', 'project.id = project_province.project_id')
-            ->where(['project.draft' => 'No'])
-            ->groupBy(['project_province.project_id'])
-            ->createCommand()->getRawSql();
-
-            $citymunTitles = ProjectCitymun::find()
-            ->select(['project_id', 'GROUP_CONCAT(DISTINCT concat(tblcitymun.citymun_m,",",tblprovince.province_m) ORDER BY tblcitymun.citymun_m ASC, tblprovince.province_m ASC SEPARATOR ", ") as title'])
-            ->leftJoin('tblcitymun', 'tblcitymun.province_c = project_citymun.province_id and tblcitymun.citymun_c = project_citymun.citymun_id')
-            ->leftJoin('tblprovince', 'tblprovince.province_c = tblcitymun.province_c')
-            ->leftJoin('project', 'project.id = project_citymun.project_id')
-            ->where(['project.draft' => 'No'])
-            ->groupBy(['project_citymun.project_id'])
-            ->createCommand()->getRawSql();
-
-            $barangayTitles = ProjectBarangay::find()
-            ->select(['project_id', 'GROUP_CONCAT(DISTINCT concat(tblbarangay.barangay_m,",",tblcitymun.citymun_m,",",tblprovince.province_m) ORDER BY tblbarangay.barangay_m ASC, tblcitymun.citymun_m ASC, tblprovince.province_m ASC SEPARATOR ", ") as title'])
-            ->leftJoin('tblbarangay', 'tblbarangay.province_c = project_barangay.province_id and tblbarangay.citymun_c = project_barangay.citymun_id and tblbarangay.barangay_c = project_barangay.barangay_id')
-            ->leftJoin('tblcitymun', 'tblcitymun.province_c = project_barangay.province_id and tblcitymun.citymun_c = project_barangay.citymun_id')
-            ->leftJoin('tblprovince', 'tblprovince.province_c = tblcitymun.province_c')
-            ->leftJoin('project', 'project.id = project_barangay.project_id')
-            ->where(['project.draft' => 'No'])
-            ->groupBy(['project_barangay.project_id'])
-            ->createCommand()->getRawSql();
-
-        $problems = ProjectProblem::find()
-                    ->select([
-                        'project_problem.quarter',
-                        'project_problem.year',
-                        'project_problem.project_id',
-                        'project_problem.nature',
-                        'project_problem.detail',
-                        'project_problem.strategy',
-                        'project_problem.responsible_entity',
-                        'project_problem.lesson_learned',
-                        'project.title as projectTitle',
-                        'sector.title as sectorTitle',
-                        'sub_sector.title as subSectorTitle',
-                        'agency.code as agencyCode',
-                        'IF(barangayTitles.title is null, IF(citymunTitles.title is null, IF(provinceTitles.title is null, IF(regionTitles.title is null, "No location", regionTitles.title), provinceTitles.title), citymunTitles.title), barangayTitles.title) as locationTitle',
-                        'COALESCE('.$financialTotal.', 0) as totalCost'
-                    ])
-                    ->leftJoin('project','project.id= project_problem.project_id')
-                    ->leftJoin('sector','sector.id= project.sector_id')
-                    ->leftJoin('sub_sector','sub_sector.id= project.sub_sector_id')
-                    ->leftJoin('agency','agency.id= project.agency_id')
-                    ->leftJoin(['regionTitles' => '('.$regionTitles.')'], 'regionTitles.project_id = project.id')
-                    ->leftJoin(['provinceTitles' => '('.$provinceTitles.')'], 'provinceTitles.project_id = project.id')
-                    ->leftJoin(['citymunTitles' => '('.$citymunTitles.')'], 'citymunTitles.project_id = project.id')
-                    ->leftJoin(['barangayTitles' => '('.$barangayTitles.')'], 'barangayTitles.project_id = project.id') 
-                    ;
-
-        $problems = $problems->leftJoin(['financials' => '('.$financials.')'], 'financials.project_id = project_problem.project_id');
-
-        if($model['year'] != '')
-        {
-            $problems = $problems->andWhere(['project_problem.year' => $model['year']]);
-        }
-        if($model['quarter'] != '')
-        {
-            $problems = $problems->andWhere(['project_problem.quarter' => $model['quarter']]);
-        }
-
-        $problems = $problems->orderBy(['project_problem.nature' => SORT_ASC])->asArray()->all();
-
-        return $this->renderAjax('form-eleven', [
-            'model' => $model,
-            'type' => 'print',
-            'problems' => $problems
-        ]);
-    }
-    public function actionDownloadFormEleven($type, $year, $quarter, $model)
-    {
-        $model = json_decode($model, true); 
-        $model['year'] = $year;
-        $model['quarter'] = $quarter;
-
-        $financials = ProjectTarget::find()->where(['target_type' => 'Financial', 'year' => $model['year']])->createCommand()->getRawSql();
-        
-        $financialTotal = 'IF(project.data_type = "Cumulative",
-        IF(COALESCE(financials.q4, 0) <= 0,
-            IF(COALESCE(financials.q3, 0) <= 0,
-                IF(COALESCE(financials.q2, 0) <= 0,
-                     COALESCE(financials.q1, 0)
-                        , COALESCE(financials.q2, 0)
-                        )
-                    , COALESCE(financials.q3, 0)
-                    )
-                , COALESCE(financials.q4, 0))
-            ,   
-            COALESCE(financials.q1, 0) +
-            COALESCE(financials.q2, 0) +
-            COALESCE(financials.q3, 0) +
-            COALESCE(financials.q4, 0)
-            )';
-
-        $regionTitles = ProjectRegion::find()
-            ->select(['project_id', 'GROUP_CONCAT(DISTINCT tblregion.abbreviation ORDER BY tblregion.abbreviation ASC SEPARATOR ", ") as title'])
-            ->leftJoin('tblregion', 'tblregion.region_c = project_region.region_id')
-            ->leftJoin('project', 'project.id = project_region.project_id')
-            ->where(['project.draft' => 'No'])
-            ->groupBy(['project_region.project_id'])
-            ->createCommand()->getRawSql();
-
-        $provinceTitles = ProjectProvince::find()
-            ->select(['project_id', 'GROUP_CONCAT(DISTINCT tblprovince.province_m ORDER BY tblprovince.province_m ASC SEPARATOR ", ") as title'])
-            ->leftJoin('tblprovince', 'tblprovince.province_c = project_province.province_id')
-            ->leftJoin('project', 'project.id = project_province.project_id')
-            ->where(['project.draft' => 'No'])
-            ->groupBy(['project_province.project_id'])
-            ->createCommand()->getRawSql();
-
-        $citymunTitles = ProjectCitymun::find()
-            ->select(['project_id', 'GROUP_CONCAT(DISTINCT concat(tblcitymun.citymun_m,",",tblprovince.province_m) ORDER BY tblcitymun.citymun_m ASC, tblprovince.province_m ASC SEPARATOR ", ") as title'])
-            ->leftJoin('tblcitymun', 'tblcitymun.province_c = project_citymun.province_id and tblcitymun.citymun_c = project_citymun.citymun_id')
-            ->leftJoin('tblprovince', 'tblprovince.province_c = tblcitymun.province_c')
-            ->leftJoin('project', 'project.id = project_citymun.project_id')
-            ->where(['project.draft' => 'No'])
-            ->groupBy(['project_citymun.project_id'])
-            ->createCommand()->getRawSql();
-
-        $barangayTitles = ProjectBarangay::find()
-            ->select(['project_id', 'GROUP_CONCAT(DISTINCT concat(tblbarangay.barangay_m,",",tblcitymun.citymun_m,",",tblprovince.province_m) ORDER BY tblbarangay.barangay_m ASC, tblcitymun.citymun_m ASC, tblprovince.province_m ASC SEPARATOR ", ") as title'])
-            ->leftJoin('tblbarangay', 'tblbarangay.province_c = project_barangay.province_id and tblbarangay.citymun_c = project_barangay.citymun_id and tblbarangay.barangay_c = project_barangay.barangay_id')
-            ->leftJoin('tblcitymun', 'tblcitymun.province_c = project_barangay.province_id and tblcitymun.citymun_c = project_barangay.citymun_id')
-            ->leftJoin('tblprovince', 'tblprovince.province_c = tblcitymun.province_c')
-            ->leftJoin('project', 'project.id = project_barangay.project_id')
-            ->where(['project.draft' => 'No'])
-            ->groupBy(['project_barangay.project_id'])
-            ->createCommand()->getRawSql();
-
-        $problems = ProjectProblem::find()
-                    ->select([
-                        'project_problem.quarter',
-                        'project_problem.year',
-                        'project_problem.project_id',
-                        'project_problem.nature',
-                        'project_problem.detail',
-                        'project_problem.strategy',
-                        'project_problem.responsible_entity',
-                        'project_problem.lesson_learned',
-                        'project.title as projectTitle',
-                        'sector.title as sectorTitle',
-                        'sub_sector.title as subSectorTitle',
-                        'agency.code as agencyCode',
-                        'IF(barangayTitles.title is null, IF(citymunTitles.title is null, IF(provinceTitles.title is null, IF(regionTitles.title is null, "No location", regionTitles.title), provinceTitles.title), citymunTitles.title), barangayTitles.title) as locationTitle',
-                        'COALESCE('.$financialTotal.', 0) as totalCost'
-                    ])
-                    ->leftJoin('project','project.id= project_problem.project_id')
-                    ->leftJoin('sector','sector.id= project.sector_id')
-                    ->leftJoin('sub_sector','sub_sector.id= project.sub_sector_id')
-                    ->leftJoin('agency','agency.id= project.agency_id')
-                    ->leftJoin(['regionTitles' => '('.$regionTitles.')'], 'regionTitles.project_id = project.id')
-                    ->leftJoin(['provinceTitles' => '('.$provinceTitles.')'], 'provinceTitles.project_id = project.id')
-                    ->leftJoin(['citymunTitles' => '('.$citymunTitles.')'], 'citymunTitles.project_id = project.id')
-                    ->leftJoin(['barangayTitles' => '('.$barangayTitles.')'], 'barangayTitles.project_id = project.id') 
-                    ;
-
-        $problems = $problems->leftJoin(['financials' => '('.$financials.')'], 'financials.project_id = project_problem.project_id');
-
-        if($model['year'] != '')
-        {
-            $problems = $problems->andWhere(['project_problem.year' => $model['year']]);
-        }
-        if($model['quarter'] != '')
-        {
-            $problems = $problems->andWhere(['project_problem.quarter' => $model['quarter']]);
-        }
-
-        $problems = $problems->orderBy(['project_problem.nature' => SORT_ASC])->asArray()->all();
-
-        $filename = 'RPMES Form 11: LIST OF PROJECT PROBLEMS/ISSUES';
-
-        if($type == 'excel')
-        {
-            header("Content-type: application/vnd.ms-excel");
-            header("Content-Disposition: attachment; filename=".$filename.".xls");
-            return $this->renderPartial('form-eleven', [
-                'model' => $model,
-                'type' => $type,
-                'problems' => $problems,
-            ]);
-        }else if($type == 'pdf')
-        {
-            $content = $this->renderPartial('form-eleven', [
-                'model' => $model,
-                'type' => $type,
-                'problems' => $problems,
-            ]);
-
-            $pdf = new Pdf([
-                'mode' => Pdf::MODE_CORE,
-                'format' => Pdf::FORMAT_LEGAL, 
-                'orientation' => Pdf::ORIENT_LANDSCAPE, 
-                'destination' => Pdf::DEST_DOWNLOAD, 
-                'filename' => $filename.'.pdf', 
-                'content' => $content,  
-                'marginLeft' => 11.4,
-                'marginRight' => 11.4,
-                'cssInline' => '*{font-family: "Arial";}
-                                table{
-                                    font-family: "Arial";
-                                    border-collapse: collapse;
-                                }
-                                thead{
-                                    font-size: 12px;
-                                    text-align: center;
-                                }
-                            
-                                td{
-                                    font-size: 10px;
-                                    border: 1px solid black;
-                                }
-                            
-                                th{
-                                    text-align: center;
-                                    border: 1px solid black;
-                                }
-                                h1,h2,h3,h4,h5,h6{
-                                    text-align: center;
-                                    font-weight: bolder;
-                                }', 
-                ]);
-        
-                $response = Yii::$app->response;
-                $response->format = \yii\web\Response::FORMAT_RAW;
-                $headers = Yii::$app->response->headers;
-                $headers->add('Content-Type', 'application/pdf');
-                return $pdf->render();
-        }
     }
 }
